backend/requirements.txt
asgiref==3.8.1
astroid==3.3.9
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dill==0.3.9
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
